Async improvements
Rewriting asynchronous routines. vfs5011.c almost works. aes* is broken.
I have a vfs5011 and an aes2501 so that is what I will concentrate on.
There is now a real thread! I've tried to mess as little at possible with
the code. Unfortunately it's in a bad shape so a big rewrite is in place.
Much work needs to be moved from the thread of libusb (callbacks) into the
main loop and made sure everything is done in order. It would probably have
been easier to keep the library synchronous and making an asynchronous
wrapper but the work has already begun, no stopping now!

Signed-off-by: Per-Ola Gustavsson <pelle@marsba.se>